python_client/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cnocr==2.3.0.1
cnstd==1.2.3.5
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
Polygon3==3.0.9.1
protobuf==4.25.1
psutil==5.9.7
pyclipper==1.3.0.post5
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.1
sentry-sdk==1.39.1
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sympy==1.12
torch==2.1.2
torchmetrics==1.2.1
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.7
urllib3==2.1.0
wandb==0.16.1
yarl==1.9.4
